Rename test file to match what it tests.
